[dy] Parse db connection url
